+++ packages/app/cypress/e2e/specs_list_e2e.cy.ts
@@ -1,4 +1,5 @@
 import { getPathForPlatform } from '../../src/paths'
+import path from 'path'
 
 describe('App: Spec List (E2E)', () => {
   const launchApp = (specFilter?: string) => {
@@ -249,6 +250,66 @@ describe('App: Spec List (E2E)', () => {
         // On Windows, when a user types `e2e/accounts`, it should match `e2e\accounts`
         clearSearchAndType('e2e/accounts')
         cy.findAllByTestId('spec-item').should('have.length', 2)
+
+        cy.findByText('No specs matched your search:').should('not.be.visible')
+      })
+
+      it('searches specs with "-" or "_" when search contains space', function () {
+        clearSearchAndType('accounts list')
+
+        cy.findAllByTestId('spec-item')
+        .should('have.length', 1)
+        .and('contain', 'accounts_list.spec.js')
+
+        cy.findByText('No specs matched your search:').should('not.be.visible')
+      })
+
+      it('searches specs with "-" or "_" when search contains "-"', function () {
+        clearSearchAndType('accounts-list')
+
+        cy.findAllByTestId('spec-item')
+        .should('have.length', 1)
+        .and('contain', 'accounts_list.spec.js')
+
+        cy.findByText('No specs matched your search:').should('not.be.visible')
+      })
+
+      it('searches specs with "-" or "_" when search contains "_"', function () {
+        clearSearchAndType('accounts_list')
+
+        cy.findAllByTestId('spec-item')
+        .should('have.length', 1)
+        .and('contain', 'accounts_list.spec.js')
+
+        cy.findByText('No specs matched your search:').should('not.be.visible')
+      })
+
+      it('searches folders with "-" or "_" when search contains space', function () {
+        clearSearchAndType('a b c')
+
+        cy.findAllByTestId('spec-list-directory')
+        .should('have.length', 1)
+        .and('contain', path.join('cypress', 'e2e', 'a-b_c'))
+
+        cy.findByText('No specs matched your search:').should('not.be.visible')
+      })
+
+      it('searches folders with "-" or "_" when search contains "-"', function () {
+        clearSearchAndType('a-b-c')
+
+        cy.findAllByTestId('spec-list-directory')
+        .should('have.length', 1)
+        .and('contain', path.join('cypress', 'e2e', 'a-b_c'))
+
+        cy.findByText('No specs matched your search:').should('not.be.visible')
+      })
+
+      it('searches folders with "-" or "_" when search contains "_"', function () {
+        clearSearchAndType('a_b_c')
+
+        cy.findAllByTestId('spec-list-directory')
+        .should('have.length', 1)
+        .and('contain', path.join('cypress', 'e2e', 'a-b_c'))
 
         cy.findByText('No specs matched your search:').should('not.be.visible')
       })